change f32 to f64 for precision

+++ src/data_point.rs
@@ -1,13 +1,13 @@
 // Data that the network can be trained on
 pub struct DataPoint {
-    pub inputs: Vec<f32>,
+    pub inputs: Vec<f64>,
     pub expected_outputs: Vec<usize>
 }
 
 impl DataPoint {
 
     // Create new DataPoint
-    pub fn new(inputs: Vec<f32>, expected_outputs:Vec<usize>) -> Self {
+    pub fn new(inputs: Vec<f64>, expected_outputs:Vec<usize>) -> Self {
         DataPoint { inputs, expected_outputs }
     }
 }
