+++ src/data_set.rs
@@ -8,8 +8,8 @@ pub struct DataSet {
 impl DataSet {
 
     // Input data must be random
-    pub fn new (mut data: Vec<DataPoint>, test_pct: f32) -> Self {
-        let split_index = (data.len() as f32 * test_pct).ceil() as usize;
+    pub fn new (mut data: Vec<DataPoint>, test_pct: f64) -> Self {
+        let split_index = (data.len() as f64 * test_pct).ceil() as usize;
         let train = data.split_off(split_index);
 
         DataSet { train, test: data }
